added next apis and deployed
